requirements.txt
streamlit>=1.40.0
boto3>=1.35.0
Pillow>=10.4.0
opencv-python-headless>=4.10.0.84
numpy>=1.26.4
python-dotenv>=1.0.1
mediapipe>=0.10.21
